+++ src/azure-cli/azure/cli/command_modules/acs/_consts.py
@@ -1,0 +1,7 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+CONST_OUTBOUND_TYPE_LOAD_BALANCER = "loadBalancer"
+CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING = "userDefinedRouting"